feat: support custom profiles through local version.json
